feat(package): setup standard dart_dev tooling
