+++ .travis.yml
@@ -2,5 +2,11 @@ language: dart
 sudo: false
 dart:
   - stable
-  - dev
-script: ./tool/travis.sh
+script:
+  - pub get --packages-dir
+  - pub run dependency_validator --ignore dart_style
+  - pub run dart_dev format --check
+  - pub run dart_dev analyze
+  - pub run dart_dev test
+  - pub run dart_dev coverage --no-html
+  - bash <(curl -s https://codecov.io/bash) -f coverage/coverage.lcov
