+++ pubspec.yaml
@@ -12,11 +12,13 @@ environment:
   sdk: ">=1.8.2 <2.0.0"
 
 dependencies:
-  path: "^1.3.0"
-  logging: ">=0.9.3<0.12.0"
   args: ">=0.11.0 <2.0.0"
+  logging: ">=0.9.3<0.12.0"
+  path: "^1.3.0"
 
 dev_dependencies:
-  test: "^0.12.13"
-  yaml: "^2.1.0"
   coverage: ">=0.7.6"
+  dart_dev: any
+  dart_style: any
+  dependency_validator: ^1.0.0
+  test: "^0.12.13"
