+++ test/unit/vm/generated_runner_test.dart
@@ -1,0 +1,13 @@
+@TestOn('vm')
+library test.unit.vm.generated_runner_test;
+
+// Generated by `pub run dart_dev gen-test-runner -d test/unit/vm -e Environment.vm --no-genHtml`
+
+import './json_schema/invalid_schemas_test.dart' as json_schema_invalid_schemas_test;
+import './json_schema/validation_test.dart' as json_schema_validation_test;
+import 'package:test/test.dart';
+
+void main() {
+  json_schema_invalid_schemas_test.main();
+  json_schema_validation_test.main();
+}
